# [Babel 8] perf: Improve traverse performance