--- a/packages/babel-traverse/test/modification.js
+++ b/packages/babel-traverse/test/modification.js
@@ -346,6 +346,67 @@ describe("modification", function () {
           "var _temp;\nexport default (_temp = fn(), x, _temp);",
         );
       });
+
+      it("edge case for babel-plugin-jest-hoist", function () {
+        let varsHoistPoint;
+        let i = 0;
+        const logs = [[], []];
+        traverse(parse(`{a;}`), {
+          BlockStatement(path) {
+            [varsHoistPoint] = path.unshiftContainer(
+              "body",
+              t.emptyStatement(),
+            );
+            path.traverse({
+              Identifier(path) {
+                if (i++ >= 5) {
+                  return;
+                }
+                logs[0].push(path.node.name);
+                varsHoistPoint.insertBefore(t.identifier("b"));
+              },
+            });
+          },
+        });
+        i = 0;
+        traverse(parse(`{a;}`), {
+          BlockStatement(path) {
+            [varsHoistPoint] = path.unshiftContainer(
+              "body",
+              t.emptyStatement(),
+            );
+            path.traverse({
+              enter() {},
+              Identifier(path) {
+                if (i++ >= 5) {
+                  return;
+                }
+                logs[1].push(path.node.name);
+                varsHoistPoint.insertBefore(t.identifier("b"));
+              },
+            });
+          },
+        });
+
+        expect(logs).toMatchInlineSnapshot(`
+          Array [
+            Array [
+              "a",
+              "b",
+              "b",
+              "b",
+              "b",
+            ],
+            Array [
+              "a",
+              "b",
+              "b",
+              "b",
+              "b",
+            ],
+          ]
+        `);
+      });
     });
   });
 });