<!--
Before making a PR, please read our contributing guidelines
https://github.com/babel/babel/blob/main/CONTRIBUTING.md

Please note that the Babel Team requires two approvals before merging most PRs.

For issue references: Add a comma-separated list of a [closing word](https://help.github.com/articles/closing-issues-via-commit-messages/) followed by the ticket number fixed by the PR. (it should be underlined in the preview if done correctly)

If you are making a change that should have a docs update: submit another PR to https://github.com/babel/website
-->

| Q                        | A <!--(Can use an emoji 👍) -->
| ------------------------ | ---
| Fixed Issues?            | `Fixes #1, Fixes #2` <!-- remove the (`) quotes and write "Fixes" before the number to link the issues -->
| Patch: Bug Fix?          |
| Major: Breaking Change?  |
| Minor: New Feature?      |
| Tests Added + Pass?      | Yes
| Documentation PR Link    | <!-- If only readme change, add `[skip ci]` to your commits -->
| Any Dependency Changes?  |
| License                  | MIT

<!-- Describe your changes below in as much detail as possible -->
This PR is a bit weird because I can't fully explain the benchmark results. Discussions are welcome!

Overall time reduction of 10%-20, with one benchmark regression (~15%) that can be fixed by merging #16964.
Also reduced reliance on the global cache and almost no global cache in Babel 8.

There are two changes in this PR
1. Organize traversal related logic into one file
2. Store cache in `parentPath`
Each of them alone gives little or no improvement, but together the performance improvement is much larger.

Even weirder is that there is a benchmark `mjs-cjs/ts-checker` that drops by more than 15%, and it has the same drop when only the second change is made, while `js-node6/typescript-5.6.2` (larger file) and `mjs-cjs/ts-parser` (similar benchmark) do not have the same situation.

I recently found that when I merged https://github.com/babel/babel/pull/16964, the situation of 1+1>2 appeared again, and the above issue disappeared.
This is confusing to me, because in my assumption the performance regression comes from recreating part of the `NodePath` (since there is no global cache anymore, when a node is moved to its parent, all related `NodePath` need to be recreated), and that PR just avoids multiple traversals and shouldn't fix it. (Traversals are faster, so the impact of the changes in #16964 in this PR should be less than 3%)

Note: This PR may be risky to merge in Babel 7, but since it doesn't require any changes to our plugins, it might be worth a try.

with #16964
```
PS F:\babel\benchmark> node --expose-gc .\all\real-case-ts-mjs.mjs
all/real-case-ts-mjs.mjs babel-parser-express.ts @ current: 51.15 ops/sec ±1.61% 256 runs (20ms)
all/real-case-ts-mjs.mjs babel-parser-express.ts @ baseline: 40.84 ops/sec ±2.95% 205 runs (24ms)
all/real-case-ts-mjs.mjs ts-parser.ts @ current: 9.41 ops/sec ±2.02% 48 runs (106ms)
all/real-case-ts-mjs.mjs ts-parser.ts @ baseline: 6.82 ops/sec ±2.78% 35 runs (147ms)
PS F:\babel\benchmark> node --expose-gc .\all\real-case-mjs-cjs.mjs
all/real-case-mjs-cjs.mjs ts-checker.mjs @ current: 1.25 ops/sec ±1.91% 10 runs (801ms)
all/real-case-mjs-cjs.mjs ts-checker.mjs @ baseline: 1.09 ops/sec ±3.99% 10 runs (922ms)
all/real-case-mjs-cjs.mjs ts-parser.mjs @ current: 10.07 ops/sec ±2.11% 51 runs (99ms)
all/real-case-mjs-cjs.mjs ts-parser.mjs @ baseline: 8.96 ops/sec ±3.3% 45 runs (112ms)
PS F:\babel\benchmark> node --expose-gc .\all\real-case-js-node6.mjs
all/real-case-js-node6.mjs jquery-3.6.js @ current: 10.35 ops/sec ±2.04% 52 runs (97ms)
all/real-case-js-node6.mjs jquery-3.6.js @ baseline: 9.21 ops/sec ±2.6% 47 runs (109ms)
all/real-case-js-node6.mjs typescript-5.6.2.js @ current: 0.34 ops/sec ±1.14% 10 runs (2970ms)
all/real-case-js-node6.mjs typescript-5.6.2.js @ baseline: 0.28 ops/sec ±4.03% 10 runs (3602ms)
PS F:\babel\benchmark> node --expose-gc .\babel-traverse\real-case.mjs
babel-traverse/real-case.mjs babel-parser-express.ts @ current: 95.92 ops/sec ±1.33% 480 runs (10ms)
babel-traverse/real-case.mjs babel-parser-express.ts @ baseline: 85.51 ops/sec ±0.76% 428 runs (12ms)
babel-traverse/real-case.mjs jquery-3.6.js @ current: 28.58 ops/sec ±1.58% 143 runs (35ms)
babel-traverse/real-case.mjs jquery-3.6.js @ baseline: 22.53 ops/sec ±1.23% 113 runs (44ms)
babel-traverse/real-case.mjs ts-checker.mjs @ current: 4.05 ops/sec ±2.13% 21 runs (247ms)
babel-traverse/real-case.mjs ts-checker.mjs @ baseline: 3.05 ops/sec ±2.34% 16 runs (328ms)
babel-traverse/real-case.mjs ts-parser.mjs @ current: 26.39 ops/sec ±1.64% 132 runs (38ms)
babel-traverse/real-case.mjs ts-parser.mjs @ baseline: 20.6 ops/sec ±1.79% 103 runs (49ms)
babel-traverse/real-case.mjs ts-parser.ts @ current: 21.94 ops/sec ±1.88% 110 runs (46ms)
babel-traverse/real-case.mjs ts-parser.ts @ baseline: 16.91 ops/sec ±1.37% 85 runs (59ms)
babel-traverse/real-case.mjs typescript-5.6.2.js @ current: 0.99 ops/sec ±3.47% 10 runs (1012ms)
babel-traverse/real-case.mjs typescript-5.6.2.js @ baseline: 0.73 ops/sec ±2.19% 10 runs (1369ms)
```

without #16964
```
PS F:\babel\benchmark> node --expose-gc .\all\real-case-ts-mjs.mjs    
all/real-case-ts-mjs.mjs babel-parser-express.ts @ current: 51.02 ops/sec ±1.75% 256 runs (20ms)
all/real-case-ts-mjs.mjs babel-parser-express.ts @ baseline: 40.44 ops/sec ±3.88% 203 runs (25ms)
all/real-case-ts-mjs.mjs ts-parser.ts @ current: 7.45 ops/sec ±1.81% 38 runs (134ms)
all/real-case-ts-mjs.mjs ts-parser.ts @ baseline: 6.85 ops/sec ±3.06% 35 runs (146ms)
PS F:\babel\benchmark> node --expose-gc .\all\real-case-mjs-cjs.mjs
all/real-case-mjs-cjs.mjs ts-checker.mjs @ current: 0.93 ops/sec ±2.63% 10 runs (1072ms)
all/real-case-mjs-cjs.mjs ts-checker.mjs @ baseline: 1.13 ops/sec ±3.04% 10 runs (884ms)
all/real-case-mjs-cjs.mjs ts-parser.mjs @ current: 10.44 ops/sec ±1.89% 53 runs (96ms)
all/real-case-mjs-cjs.mjs ts-parser.mjs @ baseline: 9.14 ops/sec ±3.65% 47 runs (109ms)
PS F:\babel\benchmark> node --expose-gc .\all\real-case-js-node6.mjs  
all/real-case-js-node6.mjs jquery-3.6.js @ current: 10.19 ops/sec ±1.83% 51 runs (98ms)
all/real-case-js-node6.mjs jquery-3.6.js @ baseline: 9.08 ops/sec ±2.76% 46 runs (110ms)
all/real-case-js-node6.mjs typescript-5.6.2.js @ current: 0.33 ops/sec ±1.52% 10 runs (3007ms)
all/real-case-js-node6.mjs typescript-5.6.2.js @ baseline: 0.27 ops/sec ±3.89% 10 runs (3687ms)
PS F:\babel\benchmark> node --expose-gc .\babel-traverse\real-case.mjs
babel-traverse/real-case.mjs babel-parser-express.ts @ current: 97.12 ops/sec ±1.24% 486 runs (10ms)
babel-traverse/real-case.mjs babel-parser-express.ts @ baseline: 84.19 ops/sec ±0.74% 421 runs (12ms)
babel-traverse/real-case.mjs jquery-3.6.js @ current: 28.25 ops/sec ±1.42% 142 runs (35ms)
babel-traverse/real-case.mjs jquery-3.6.js @ baseline: 22.43 ops/sec ±1.37% 113 runs (45ms)
babel-traverse/real-case.mjs ts-checker.mjs @ current: 4.05 ops/sec ±1.74% 21 runs (247ms)
babel-traverse/real-case.mjs ts-checker.mjs @ baseline: 3.07 ops/sec ±2.26% 16 runs (325ms)
babel-traverse/real-case.mjs ts-parser.mjs @ current: 26.19 ops/sec ±1.52% 132 runs (38ms)
babel-traverse/real-case.mjs ts-parser.mjs @ baseline: 20.55 ops/sec ±1.29% 103 runs (49ms)
babel-traverse/real-case.mjs ts-parser.ts @ current: 21.85 ops/sec ±1.67% 110 runs (46ms)
babel-traverse/real-case.mjs ts-parser.ts @ baseline: 16.54 ops/sec ±1.35% 83 runs (60ms)
babel-traverse/real-case.mjs typescript-5.6.2.js @ current: 0.99 ops/sec ±3.61% 10 runs (1014ms)
babel-traverse/real-case.mjs typescript-5.6.2.js @ baseline: 0.71 ops/sec ±3.9% 10 runs (1404ms)
```