@liuxingbaoyu 2025-02-17
To avoid possible regressions, I moved it to Babel 8.